# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "cliagents API",
    "description": "REST API for AI agent orchestration. Supports multiple AI CLI tools (Claude Code, Gemini CLI, Codex CLI, etc.) via a unified HTTP interface.",
    "version": "1.0.0",
    "contact": {
      "name": "cliagents"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/sessions": {
      "post": {
        "summary": "Create a new session",
        "description": "Create a new agent session with the specified adapter",
        "operationId": "createSession",
        "tags": ["Sessions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "404": {
            "description": "Adapter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Adapter CLI not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all sessions",
        "description": "Get a list of all active sessions",
        "operationId": "listSessions",
        "tags": ["Sessions"],
        "responses": {
          "200": {
            "description": "List of sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionInfo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}": {
      "get": {
        "summary": "Get session info",
        "description": "Get detailed information about a specific session",
        "operationId": "getSession",
        "tags": ["Sessions"],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Terminate session",
        "description": "Terminate and clean up a session",
        "operationId": "terminateSession",
        "tags": ["Sessions"],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session terminated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "terminated"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/models/{model}": {
      "get": {
        "summary": "Get model details",
        "description": "Returns details about a specific model. OpenAI-compatible endpoint.",
        "operationId": "getModel",
        "tags": ["OpenAI Compatible"],
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Model ID (e.g., claude-sonnet-4-20250514, gpt-4o, gemini-2.5-flash)"
          }
        ],
        "responses": {
          "200": {
            "description": "Model details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIModel"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Create chat completion",
        "description": "Creates a chat completion using an OpenAI-compatible request format. Routes to the appropriate CLI adapter based on the model name. Supports both streaming and non-streaming modes.",
        "operationId": "createChatCompletion",
        "tags": ["OpenAI Compatible"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response (or SSE stream if stream=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream in OpenAI format"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIError"
                }
              }
            }
          },
          "503": {
            "description": "Model/CLI not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "adapter": {
            "type": "string",
            "description": "Adapter to use",
            "example": "claude-code",
            "enum": ["claude-code", "gemini-cli", "codex-cli", "aider", "goose", "amazon-q", "plandex", "continue-cli", "mistral-vibe", "shell-gpt", "aichat", "github-copilot"]
          },
          "model": {
            "type": "string",
            "description": "Specific model to use (adapter-specific)"
          },
          "systemPrompt": {
            "type": "string",
            "description": "System prompt to initialize the session"
          },
          "workDir": {
            "type": "string",
            "description": "Working directory for the session"
          },
          "allowedTools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of allowed tools (restricts agent capabilities)"
          },
          "jsonSchema": {
            "type": "object",
            "description": "JSON Schema for structured output (Claude only)"
          },
          "temperature": {
            "type": "number",
            "description": "Sampling temperature 0.0-2.0 (Gemini only)"
          },
          "top_p": {
            "type": "number",
            "description": "Nucleus sampling threshold 0.0-1.0 (Gemini only)"
          },
          "top_k": {
            "type": "integer",
            "description": "Top-k sampling (Gemini only)"
          },
          "max_output_tokens": {
            "type": "integer",
            "description": "Maximum output tokens"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "adapter": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["ready", "running", "stable", "error"]
          },
          "model": {
            "type": "string"
          }
        }
      },
      "SessionInfo": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "adapterName": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "description": "Unix timestamp in milliseconds"
          },
          "lastActivity": {
            "type": "integer",
            "description": "Unix timestamp in milliseconds"
          },
          "status": {
            "type": "string",
            "enum": ["stable", "running", "error"]
          },
          "messageCount": {
            "type": "integer"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code",
                "enum": ["invalid_request_error", "missing_parameter", "session_not_found", "adapter_not_found", "adapter_unavailable", "timeout_error", "cli_error", "internal_error"]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "type": {
                "type": "string"
              },
              "param": {
                "type": "string",
                "description": "Parameter that caused the error (if applicable)"
              }
            }
          }
        }
      },
      "OpenAIModel": {
        "type": "object",
        "description": "OpenAI-compatible model object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Model identifier",
            "example": "claude-sonnet-4-20250514"
          },
          "object": {
            "type": "string",
            "example": "model"
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp of model creation"
          },
          "owned_by": {
            "type": "string",
            "description": "The adapter that provides this model",
            "example": "claude-code"
          },
          "available": {
            "type": "boolean",
            "description": "Whether the model's CLI is available"
          }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "description": "OpenAI-compatible chat completion request",
        "required": ["model", "messages"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to use. Maps to CLI adapters: gpt-* → codex-cli, claude-* → claude-code, gemini-* → gemini-cli",
            "example": "claude-sonnet-4-20250514"
          },
          "messages": {
            "type": "array",
            "description": "List of messages in the conversation",
            "items": {
              "type": "object",
              "required": ["role", "content"],
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["system", "user", "assistant"],
                  "description": "Role of the message author"
                },
                "content": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Message content as text"
                    },
                    {
                      "type": "array",
                      "description": "Message content as array (for vision)",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": ["text", "image_url"]
                          },
                          "text": {
                            "type": "string"
                          },
                          "image_url": {
                            "type": "object",
                            "properties": {
                              "url": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "stream": {
            "type": "boolean",
            "default": false,
            "description": "If true, returns SSE stream"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "description": "Sampling temperature (Gemini only)"
          },
          "max_tokens": {
            "type": "integer",
            "description": "Maximum tokens to generate"
          },
          "top_p": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Nucleus sampling (Gemini only)"
          },
          "stop": {
            "oneOf": [
              {"type": "string"},
              {"type": "array", "items": {"type": "string"}}
            ],
            "description": "Stop sequences"
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "description": "OpenAI-compatible chat completion response",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique completion ID",
            "example": "chatcmpl-abc123"
          },
          "object": {
            "type": "string",
            "example": "chat.completion"
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp"
          },
          "model": {
            "type": "string",
            "description": "Model used"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer"
                },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "example": "assistant"
                    },
                    "content": {
                      "type": "string"
                    }
                  }
                },
                "logprobs": {
                  "type": "object",
                  "nullable": true
                },
                "finish_reason": {
                  "type": "string",
                  "enum": ["stop", "length", "content_filter", "tool_calls"]
                }
              }
            }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": {
                "type": "integer"
              },
              "completion_tokens": {
                "type": "integer"
              },
              "total_tokens": {
                "type": "integer"
              }
            }
          }
        }
      },
      "OpenAIError": {
        "type": "object",
        "description": "OpenAI-compatible error response",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": ["invalid_request_error", "server_error", "internal_error"]
              },
              "param": {
                "type": "string",
                "nullable": true
              },
              "code": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      }
    }
  },
...
}